% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeChiSquaredTest}
\alias{computeChiSquaredTest}
\title{Compute chi-squared test}
\usage{
computeChiSquaredTest(Mat, Groups, classes)
}
\arguments{
\item{Mat}{Matrix of digitized binary or ternary data with each column corresponding to a sample and each 
row corresponding to a feature}

\item{Groups}{Factor indicating class association of samples}

\item{classes}{Vector of class labels; the test will be applied between the first two classes given.}
}
\description{
Given a binary or ternary data matrix with class associations of samples, computes chi-squared tests
for each feature between two given classes
}
\keyword{chi-squared}
