% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeDigitization}
\alias{computeDigitization}
\title{Compute ternary digitization}
\usage{
computeDigitization(Mat, baseMat, classes = NULL, gamma = c(1:9/100,
  1:9/10), beta = 0.95, method = "partial", par = TRUE, alpha = 0.01,
  verbose = TRUE)
}
\arguments{
\item{Mat}{Matrix of data to be digitized (usually in quantiles), with each column corresponding to a sample and each 
row corresponding to a feature}

\item{baseMat}{Matrix of baseline data (usually in quantiles), with each column corresponding to a sample and each 
row corresponding to a feature}

\item{classes}{}

\item{gamma}{Range of gamma values to search through. 
By default gamma = {0.01, 0.02, ... 0.09, 0.1, 0.2, ..., 0.9}.}

\item{beta}{Parameter for eliminating outliers (0 < beta < 1). By default beta=0.95.}

\item{method}{Method of computing distance to j'th nearest neighbour. Options are 
"partial" (default option and usually the fastest) for using partial quicksort, 
"fullsort" for using full quicksort, "kdtree" for using a kd-tree (requires the package nabor), 
"quantile_dist" for computing quantiles over pairwise distances, "ranks" for using the rank function.}

\item{par}{Logical indicating whether to compute features parallelly with mclapply (defaults to TRUE).}

\item{alpha}{Expected proportion of divergent features per sample to be estimated. The optimal gamma providing
this level of divergence in the baseline data will be searched for.}

\item{verbose}{Logical indicating whether to print status related messages during computation (defaults
to TRUE).}
}
\description{
Function for obtaining the digitized form, along with other relevant statistics and measures 
given a data matrix and a baseline matrix
}
\examples{
data(divergence)
D = computeDigitization(breastTCGA_Mat[, breastTCGA_Group=="NORMAL"], breastTCGA_Mat[, breastTCGA_Group!="NORMAL"])

}
\keyword{digitize}
\keyword{ternary}
