% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeMultivariateBinaryMatrix}
\alias{computeMultivariateBinaryMatrix}
\title{Compute binary matrix given feature set support}
\usage{
computeMultivariateBinaryMatrix(Mat, genesets, Centermatrix_list, Radius_list,
  method = "euclidean")
}
\description{
Compute binary matrix given feature set support
}
