% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeMultivariateDigitization}
\alias{computeMultivariateDigitization}
\title{Compute binary digitization for feature sets}
\usage{
computeMultivariateDigitization(Mat, basemat, genesets, gamma = 1:9/10,
  beta = 0.95, alpha = 0.01, method = "euclidean", findGamma = TRUE)
}
\description{
Compute binary digitization for feature sets
}
