% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeMultivariateSupport}
\alias{computeMultivariateSupport}
\title{Compute the support for a list of feature sets}
\usage{
computeMultivariateSupport(Mat, genesets, gamma = 1:9/10, beta = 0.95,
  method = "euclidean")
}
\description{
Compute the support for a list of feature sets
}
