% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computePercentileMatrix}
\alias{computePercentileMatrix}
\title{Compute percentile transformations}
\usage{
computePercentileMatrix(Mat, rank_samples = TRUE)
}
\arguments{
\item{Mat}{Matrix of data, with each column corresponding to a sample and each row corresponding to a feature.}
}
\value{
A matrix of the same dimensions as Mat with the percentile data.
}
\description{
Function for computing the percentile transformation for one or more samples
supplied as columns of a matrix.
}
\examples{
data(divergence)
percentileMat = computePercentileMatrix(breastTCGA_Mat)

}
