% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeQuantileMatrix}
\alias{computeQuantileMatrix}
\title{Compute quantile transformations}
\usage{
computeQuantileMatrix(Mat)
}
\arguments{
\item{Mat}{Matrix of data, with each column corresponding to a sample and each row corresponding to a feature.}
}
\value{
A matrix of the same dimensions as Mat with the quantile data.
}
\description{
Function for computing the quantile transformation for one or more samples
supplied as columns of a matrix.
}
\examples{
data(divergence)
quantileMat = computeQuantileMatrix(breastTCGA_Mat)

}
