% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeUnivariateDigitization}
\alias{computeUnivariateDigitization}
\title{Compute ternary digitization}
\usage{
computeUnivariateDigitization(Mat, baseMat, computeQuantiles = TRUE,
  gamma = c(1:9/100, 1:9/10), beta = 0.95, alpha = 0.01,
  method = "partial", parallel = TRUE, verbose = TRUE, findGamma = TRUE,
  Groups = NULL, classes = NULL)
}
\arguments{
\item{Mat}{Matrix of data to be digitized (usually in quantiles), with each column corresponding to a sample and each 
row corresponding to a feature}

\item{baseMat}{Matrix of baseline data (usually in quantiles), with each column corresponding to a sample and each 
row corresponding to a feature}

\item{computeQuantiles}{Apply quantile transformation to both data and baseline matrices (TRUE or FALSE; defaults to TRUE).}

\item{gamma}{Range of gamma values to search through. 
By default gamma = {0.01, 0.02, ... 0.09, 0.1, 0.2, ..., 0.9}.}

\item{beta}{Parameter for eliminating outliers (0 < beta < 1). By default beta=0.95.}

\item{alpha}{Expected proportion of divergent features per sample to be estimated. The optimal gamma providing
this level of divergence in the baseline data will be searched for.}

\item{method}{Method of computing distance to j'th nearest neighbour. Options are 
"partial" (default option and usually the fastest) for using partial quicksort, 
"fullsort" for using full quicksort, "kdtree" for using a kd-tree (requires the package nabor), 
"quantile_dist" for computing quantiles over pairwise distances, "ranks" for using the rank function.}

\item{parallel}{Logical indicating whether to compute features parallelly with mclapply (defaults to TRUE).}

\item{verbose}{Logical indicating whether to print status related messages during computation (defaults
to TRUE).}

\item{findGamma}{Logical indicating whether to search for optimal gamma values through the given gamma values (defaults to 
TRUE). If FALSE, the first value given in gamma will be used.}

\item{Groups}{Factor indicating class association of samples}

\item{classes}{Vector of class labels}
}
\value{
A list
}
\description{
Function for obtaining the digitized form, along with other relevant statistics and measures 
given a data matrix and a baseline matrix
}
\examples{
data(divergence)
D = computeDigitization(breastTCGA_Mat[, breastTCGA_Group=="NORMAL"], breastTCGA_Mat[, breastTCGA_Group!="NORMAL"])

}
\keyword{digitize}
\keyword{ternary}
