% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{findMultivariateGammaAndSupport}
\alias{findMultivariateGammaAndSupport}
\title{Find optimal gamma and corresponding support for list of feature sets}
\usage{
findMultivariateGammaAndSupport(Mat, genesets, gamma = 1:9/10, beta = 0.95,
  alpha = 0.01, method = "euclidean")
}
\description{
Find optimal gamma and corresponding support for list of feature sets
}
